requirements.txt
google-genai
gradio
SpeechRecognition
gTTS
pydub
python-dotenv